Initial commit, mostly setup.
After some research, decided to use commander.js for the tool. Spent
most of the time understanding how to set it up and understand how to
test it.

Some initial tests have been created for basic usage of the
tool (calling it "slint" for right now) and some functions that will
probably be required. JSDocs are present where appropriate for this
stage.